// h5/src/apis/models/ProjectConstructionApprovalItemVo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectConstructionApprovalItemVo
 */
export interface ProjectConstructionApprovalItemVo {
    /**
     * 项目类型
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    type?: string;
    /**
     * 事项名称
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    name?: string;
    /**
     * 审批部门
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    department?: string;
    /**
     * 审批结果
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    result?: string;
    /**
     * 审批时间
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    time?: string;
    /**
     * 所属步骤
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    stage?: string;
    /**
     * 是否基本事项
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    isBasic?: string;
    /**
     * 办件时限
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    limitTime?: string;
    /**
     * 事项id
     * @type {string}
     * @memberof ProjectConstructionApprovalItemVo
     */
    itemCode?: string;
}


/**
 * Check if a given object implements the ProjectConstructionApprovalItemVo interface.
 */
export function instanceOfProjectConstructionApprovalItemVo(value: object): value is ProjectConstructionApprovalItemVo {
    return true;
}

export function ProjectConstructionApprovalItemVoFromJSON(json: any): ProjectConstructionApprovalItemVo {
    return ProjectConstructionApprovalItemVoFromJSONTyped(json, false);
}

export function ProjectConstructionApprovalItemVoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectConstructionApprovalItemVo {
    if (json == null) {
        return json;
    }
    return {
        
        'type': json['type'] == null ? undefined : json['type'],
        'name': json['name'] == null ? undefined : json['name'],
        'department': json['department'] == null ? undefined : json['department'],
        'result': json['result'] == null ? undefined : json['result'],
        'time': json['time'] == null ? undefined : json['time'],
        'stage': json['stage'] == null ? undefined : json['stage'],
        'isBasic': json['isBasic'] == null ? undefined : json['isBasic'],
        'limitTime': json['limitTime'] == null ? undefined : json['limitTime'],
        'itemCode': json['itemCode'] == null ? undefined : json['itemCode'],
    };
}

export function ProjectConstructionApprovalItemVoToJSON(json: any): ProjectConstructionApprovalItemVo {
    return ProjectConstructionApprovalItemVoToJSONTyped(json, false);
}

export function ProjectConstructionApprovalItemVoToJSONTyped(value?: ProjectConstructionApprovalItemVo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': value['type'],
        'name': value['name'],
        'department': value['department'],
        'result': value['result'],
        'time': value['time'],
        'stage': value['stage'],
        'isBasic': value['isBasic'],
        'limitTime': value['limitTime'],
        'itemCode': value['itemCode'],
    };
}